requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.3.7
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
beautifulsoup4>=4.12.2
requests>=2.31.0
pypdf>=3.17.0
docx2txt>=0.8
pandas>=2.1.1
matplotlib>=3.8.0
jinja2>=3.1.2
duckduckgo-search>=3.9.3
pytest>=7.4.3
python-multipart>=0.0.6
playwright>=1.40.0
markdownify>=0.11.6
fake-useragent>=1.2.1
aiohttp>=3.8.5
aiohttp-retry>=2.8.3
pdfplumber>=0.10.2
newspaper3k>=0.2.8
pymilvus>=2.3.0
huggingface-hub>=0.19.4
FlagEmbedding>=1.1.5
Pillow>=10.1.0
markdown>=3.5.1
aiofiles>=23.2.1
PyPDF2>=3.0.1
